TINY-7688: move getAnchorSpec to Coords + fallback logic
